+++ Packages/XCharts/Runtime/Serie/Serie.cs
@@ -264,6 +264,7 @@ namespace XCharts.Runtime
         [SerializeField] private float m_BarGap = 0.1f;
         [SerializeField] private float m_BarZebraWidth = 4f;
         [SerializeField] private float m_BarZebraGap = 2f;
+        [SerializeField] [Since("v3.15.0")]private bool m_IgnoreZeroOccupy = false;
 
         [SerializeField] private float m_Min;
         [SerializeField] private float m_Max;
@@ -632,7 +633,8 @@ namespace XCharts.Runtime
             set { if (PropertyUtil.SetStruct(ref m_BarGap, value)) SetVerticesDirty(); }
         }
         /// <summary>
-        /// 斑马线的粗细。
+        /// The width of zebra bar. It is the width of each zebra stripe. When the value is 0, there is no zebra stripe.
+        /// ||斑马线的粗细。
         /// </summary>
         public float barZebraWidth
         {
@@ -640,12 +642,23 @@ namespace XCharts.Runtime
             set { if (PropertyUtil.SetStruct(ref m_BarZebraWidth, value < 0 ? 0 : value)) SetVerticesDirty(); }
         }
         /// <summary>
-        /// 斑马线的间距。
+        /// The gap of zebra bar. It is the distance between two zebra stripes. When the value is 0, there is no gap between stripes.
+        /// ||斑马线的间距。
         /// </summary>
         public float barZebraGap
         {
             get { return m_BarZebraGap; }
             set { if (PropertyUtil.SetStruct(ref m_BarZebraGap, value < 0 ? 0 : value)) SetVerticesDirty(); }
+        }
+        /// <summary>
+        /// Whether to ignore the zero value bar occupy. When enabled, the bar with zero value will not occupy space, 
+        /// and the gap between bars will be automatically adjusted according to the actual displayed bars. Generally used in bar chart.
+        /// ||柱图是否忽略值为0的柱子占位。开启后，值为0的柱子将不会占用空间，柱子之间的间距会根据实际显示的柱子自动调整。一般用在柱状图中。
+        /// </summary>
+        public bool ignoreZeroOccupy
+        {
+            get { return m_IgnoreZeroOccupy; }
+            set { if (PropertyUtil.SetStruct(ref m_IgnoreZeroOccupy, value)) SetVerticesDirty(); }
         }
 
         /// <summary>
@@ -1337,6 +1350,7 @@ namespace XCharts.Runtime
                 m_Data.RemoveAt(index);
                 m_NeedUpdateFilterData = true;
                 labelDirty = true;
+                titleDirty = true;
                 dataDirty = true;
             }
         }
@@ -1363,6 +1377,7 @@ namespace XCharts.Runtime
             SetVerticesDirty();
             CheckDataName(dataName);
             labelDirty = true;
+            titleDirty = true;
             dataDirty = true;
             return serieData;
         }
@@ -1377,6 +1392,8 @@ namespace XCharts.Runtime
             context.totalDataIndex++;
             SetVerticesDirty();
             dataDirty = true;
+            labelDirty = true;
+            titleDirty = true;
             m_NeedUpdateFilterData = true;
         }
 
@@ -1411,6 +1428,7 @@ namespace XCharts.Runtime
             SetVerticesDirty();
             CheckDataName(dataName);
             labelDirty = true;
+            titleDirty = true;
             return serieData;
         }
 
@@ -1443,6 +1461,7 @@ namespace XCharts.Runtime
             SetVerticesDirty();
             CheckDataName(dataName);
             labelDirty = true;
+            titleDirty = true;
             return serieData;
         }
 
@@ -1477,6 +1496,7 @@ namespace XCharts.Runtime
                 SetVerticesDirty();
                 CheckDataName(dataName);
                 labelDirty = true;
+                titleDirty = true;
                 return serieData;
             }
         }
@@ -1512,6 +1532,7 @@ namespace XCharts.Runtime
                 SetVerticesDirty();
                 CheckDataName(dataName);
                 labelDirty = true;
+                titleDirty = true;
                 return serieData;
             }
         }
@@ -1805,6 +1826,7 @@ namespace XCharts.Runtime
                 {
                     SetVerticesDirty();
                     dataDirty = true;
+                    titleDirty = true;
                 }
                 return flag;
             }